refactor(signup): organizar tela de cadastro em pasta própria com arquivos Signup e styles
